feat: start websocket server
