+++ package.json
@@ -3,32 +3,31 @@
   "version": "1.0.0",
   "description": "Template for the WebSocket remote control task",
   "main": "index.js",
-  "type": "module",
   "scripts": {
     "start": "node ./index.js",
-    "start:dev": "nodemon ./index.js",
+    "start:dev": "nodemon ./index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Andrei Auchynnikau",
   "keywords": [],
   "license": "ISC",
   "devDependencies": {
-    "nodemon": "^2.0.16",
-    "typescript": "^5.1.6",
     "@types/jest": "^29.5.2",
     "@types/node": "^20.3.3",
     "@types/supertest": "^2.0.12",
     "@types/uuid": "^9.0.2",
+    "@types/ws": "^8.5.5",
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
     "eslint": "^8.44.0",
+    "nodemon": "^2.0.16",
     "typescript": "^5.1.6"
   },
   "dependencies": {
-    "ws": "^8.8.0",
     "dotenv": "^16.3.1",
     "jest": "^29.6.1",
     "ts-jest": "^29.1.1",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "ws": "^8.8.0"
   }
 }
