+++ tsconfig.json
@@ -2,10 +2,10 @@
   "compilerOptions": {
     "target": "es6",
     "module": "CommonJS",
-    "rootDir": "./src",
+    // "rootDir": "./src",
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": "./src",
+    // "baseUrl": "./src",
     "strict": true,
     "esModuleInterop": true,
     "moduleResolution": "Node"
@@ -13,6 +13,6 @@
   "ts-node": {
     "esm": true
   },
-  "include": ["src/*.ts", "src/server.ts"],
+  "include": ["src/*.ts", "./index.ts"],
   "exclude": ["node_modules", "dist", "front"]
 }
